+++ src/components/ui/UserAvatar.svelte
